requirements.txt
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.2
matplotlib==3.7.5
streamlit==1.33.0
safetensor==0.4.3